@@ -13,34 +13,36 @@
         "issues": "https://github.com/DirectoryLister/DirectoryLister/issues"
     },
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^8.1 || ^8.2 || ^8.3 || ^8.4",
         "ext-dom": "*",
         "ext-fileinfo": "*",
         "ext-zip": "*",
         "erusev/parsedown-extra": "^0.8.1",
         "filp/whoops": "^2.7",
-        "phlak/splat": "^4.0",
-        "php-di/php-di": "^6.0",
+        "phlak/splat": "^5.0",
+        "maennchen/zipstream-php": "^3.1",
+        "php-di/php-di": "^7.0",
         "php-di/slim-bridge": "^3.0",
-        "psr/cache": "^1.0",
-        "psr/http-message": "^1.0.1",
+        "psr/http-message": "^2.0",
         "slim/psr7": "^1.0",
         "slim/slim": "^4.3",
         "slim/twig-view": "^3.0",
-        "symfony/cache": "^5.0",
-        "symfony/finder": "^5.0",
-        "symfony/translation": "^5.0",
-        "symfony/yaml": "^5.0",
-        "tightenco/collect": "^8.0",
+        "symfony/cache": "^6.0",
+        "symfony/finder": "^6.0",
+        "symfony/translation": "^6.0",
+        "symfony/yaml": "^6.0",
+        "tightenco/collect": "^9.52",
         "vlucas/phpdotenv": "^5.0"
     },
     "require-dev": {
+        "beyondcode/expose": "^2.0",
+        "friendsofphp/php-cs-fixer": "!=3.5.0",
         "johnkary/phpunit-speedtrap": "^4.0",
         "phlak/coding-standards": "^2.0",
-        "psy/psysh": "^0.10",
-        "symfony/var-dumper": "^5.0",
-        "vimeo/psalm": "^4.3",
-        "yoast/phpunit-polyfills": "^1.0"
+        "phpstan/phpstan": "^2.0",
+        "psy/psysh": "^0.12.0",
+        "symfony/var-dumper": "^6.0",
+        "yoast/phpunit-polyfills": "^3.0"
     },
     "suggest": {
         "ext-apcu": "Required to use the APCu cache driver",
@@ -58,8 +60,12 @@
         }
     },
     "config": {
-        "sort-packages": true,
         "optimize-autoloader": true,
+        "platform": {
+            "php": "8.1"
+        },
+        "platform-check": true,
+        "sort-packages": true,
         "vendor-dir": "app/vendor"
     },
     "scripts": {
