@@ -8,28 +8,22 @@ use Tightenco\Collect\Support\Collection;
 
 class Breadcrumbs extends ViewFunction
 {
-    /** @var string The function name */
-    protected $name = 'breadcrumbs';
-
-    /** @var Config The application configuration */
-    protected $config;
-
-    /** @var string The directory separator */
-    protected $directorySeparator;
+    protected string $name = 'breadcrumbs';
 
     /** Create a new Breadcrumbs object. */
     public function __construct(
-        Config $config,
-        string $directorySeparator = DIRECTORY_SEPARATOR
-    ) {
-        $this->config = $config;
-        $this->directorySeparator = $directorySeparator;
-    }
+        private Config $config,
+        private string $directorySeparator = DIRECTORY_SEPARATOR
+    ) {}
 
-    /** Build a collection of breadcrumbs for a given path. */
+    /**
+     * Build a collection of breadcrumbs for a given path.
+     *
+     * @return Collection<int, string>
+     */
     public function __invoke(string $path): Collection
     {
-        return Str::explode($path, $this->directorySeparator)->diff(
+        return Str::explode($path, $this->directorySeparator)->diffAssoc(
             explode($this->directorySeparator, $this->config->get('base_path'))
         )->filter(static function (string $crumb): bool {
             return ! in_array($crumb, [null, '.']);
