@@ -4,8 +4,7 @@ namespace App\ViewFunctions;
 
 class FileUrl extends Url
 {
-    /** @var string The function name */
-    protected $name = 'file_url';
+    protected string $name = 'file_url';
 
     /** Return the URL for a given path and action. */
     public function __invoke(string $path = '/'): string
