@@ -7,8 +7,7 @@ use Symfony\Component\Finder\SplFileInfo;
 
 class SizeForHumans extends ViewFunction
 {
-    /** @var string The function name */
-    protected $name = 'size_for_humans';
+    protected string $name = 'size_for_humans';
 
     /** Get the human readable file size from a file object. */
     public function __invoke(SplFileInfo $file): string
