@@ -11,7 +11,7 @@ class ViewFunctionTest extends TestCase
     public function test_it_can_be_extended(): void
     {
         $viewFunction = new class extends ViewFunction {
-            protected $name = 'foo';
+            protected string $name = 'foo';
         };
 
         $this->assertInstanceOf(ViewFunction::class, $viewFunction);
