@@ -2,11 +2,13 @@ import Alpine from 'alpinejs';
 import hljs from 'highlight.js';
 
 // Alpine.js components
-import Application from './components/application.js';
-import FileInfoModal from './components/file-info-modal.js';
+import Application from './components/Application.js';
+import FileInfoModal from './components/FileInfoModal.js';
+
+Alpine.store('theme', 'light');
 
 Alpine.data('application', Application);
-Alpine.data('file-info-modal', FileInfoModal);
+Alpine.data('fileInfoModal', FileInfoModal);
 
 // Run Alpine.js
 Alpine.start();
