@@ -1,18 +1,16 @@
 export default () => ({
-    theme: 'light',
-
     init() {
         const [cookie, theme] = document.cookie.match(/theme=(dark|light)/) || [null, null];
 
-        this.theme = theme || (window.matchMedia && window.matchMedia('(prefers-color-scheme: light)').matches ? 'light' : 'dark');
+        this.$store.theme = theme || (window.matchMedia && window.matchMedia('(prefers-color-scheme: light)').matches ? 'light' : 'dark');
 
         if (cookie === undefined) {
-            this.storeThemeCookie(this.theme);
+            this.storeThemeCookie(this.$store.theme);
         }
     },
 
     toggleTheme() {
-        this.theme = this.theme == 'light' ? 'dark' : 'light';
+        this.$store.theme = this.$store.theme == 'light' ? 'dark' : 'light';
     },
 
     storeThemeCookie(theme) {
