@@ -13,18 +13,11 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 
 class TranslationFactory
 {
-    /** @var Config The applicaiton configuration */
-    protected $config;
-
-    /** @var CacheInterface The application cache */
-    protected $cache;
-
     /** Create a new TranslationFactory object. */
-    public function __construct(Config $config, CacheInterface $cache)
-    {
-        $this->config = $config;
-        $this->cache = $cache;
-    }
+    public function __construct(
+        private Config $config,
+        private CacheInterface $cache
+    ) {}
 
     /** Initialize and return the translation component. */
     public function __invoke(): TranslatorInterface
@@ -37,7 +30,7 @@ class TranslationFactory
         }
 
         $translator = new Translator($language);
-        $translator->addLoader('yaml', new YamlFileLoader());
+        $translator->addLoader('yaml', new YamlFileLoader);
 
         foreach ($translations as $language) {
             $translator->addResource('yaml', sprintf(
