@@ -3,6 +3,7 @@
 namespace App\Factories;
 
 use App\Config;
+use App\ViewFunctions\ViewFunction;
 use Invoker\CallableResolver;
 use Slim\Views\Twig;
 use Twig\Extension\CoreExtension;
@@ -11,20 +12,11 @@ use Twig\TwigFunction;
 
 class TwigFactory
 {
-    /** @var Config The application configuration */
-    protected $config;
-
-    /** @var CallableResolver The callable resolver */
-    protected $callableResolver;
-
     /** Create a new TwigFactory object. */
     public function __construct(
-        Config $config,
-        CallableResolver $callableResolver
-    ) {
-        $this->config = $config;
-        $this->callableResolver = $callableResolver;
-    }
+        private Config $config,
+        private CallableResolver $callableResolver
+    ) {}
 
     /** Initialize and return the Twig component. */
     public function __invoke(): Twig
@@ -33,16 +25,17 @@ class TwigFactory
             $this->config->get('views_path')
         ), ['cache' => $this->config->get('view_cache')]);
 
-        $environment = $twig->getEnvironment();
-        $core = $environment->getExtension(CoreExtension::class);
+        /** @var CoreExtension $core */
+        $core = $twig->getEnvironment()->getExtension(CoreExtension::class);
 
         $core->setDateFormat($this->config->get('date_format'), '%d days');
         $core->setTimezone($this->config->get('timezone'));
 
         foreach ($this->config->get('view_functions') as $function) {
+            /** @var ViewFunction&callable $function */
             $function = $this->callableResolver->resolve($function);
 
-            $environment->addFunction(
+            $twig->getEnvironment()->addFunction(
                 new TwigFunction($function->name(), $function)
             );
         }
