@@ -11,25 +11,12 @@ use Whoops\RunInterface;
 
 class WhoopsMiddleware
 {
-    /** @var RunInterface The Whoops component */
-    protected $whoops;
-
-    /** @var PrettyPageHandler The pretty page handler */
-    protected $pageHandler;
-
-    /** @var JsonResponseHandler The JSON response handler */
-    protected $jsonHandler;
-
     /** Create a new WhoopseMiddleware object. */
     public function __construct(
-        RunInterface $whoops,
-        PrettyPageHandler $pageHandler,
-        JsonResponseHandler $jsonHandler
-    ) {
-        $this->whoops = $whoops;
-        $this->pageHandler = $pageHandler;
-        $this->jsonHandler = $jsonHandler;
-    }
+        private RunInterface $whoops,
+        private PrettyPageHandler $pageHandler,
+        private JsonResponseHandler $jsonHandler
+    ) {}
 
     /** Invoke the WhoopseMiddleware class. */
     public function __invoke(Request $request, RequestHandler $handler): ResponseInterface
@@ -40,7 +27,7 @@ class WhoopsMiddleware
 
         $this->whoops->pushHandler($this->pageHandler);
 
-        if (in_array('application/json', explode(',', $request->getHeaderLine('Accept')))) {
+        if (in_array('application/json', explode(',', (string) $request->getHeaderLine('Accept')))) {
             $this->whoops->pushHandler($this->jsonHandler);
         }
 
