@@ -7,11 +7,11 @@ use Slim\App;
 use Tests\TestCase;
 
 /** @covers \App\Bootstrap\AppManager */
-class AppManangerTest extends TestCase
+class AppManagerTest extends TestCase
 {
     public function test_it_returns_an_app_instance(): void
     {
-        $app = (new AppManager($this->container))($this->testFilesPath);
+        $app = (new AppManager($this->container))();
 
         $this->assertInstanceOf(App::class, $app);
         $this->assertSame($this->container, $app->getContainer());
